F2Player: fix float arithmetics
this caused wrong notes in
AAOCG_-_HWiNFO32_1.32kg.xm

+++ neoart/flod/fasttracker/F2Player.c
@@ -1308,9 +1308,9 @@ static int amiga(int note, int finetune) {
 	int period = PERIODS[++note];
 
 	if (finetune < 0) {
-		delta = (PERIODS[--note] - period) / 64;
+		delta = (float) (PERIODS[--note] - period) / 64.f;
 	} else if (finetune > 0) {
-		delta = (period - PERIODS[++note]) / 64;
+		delta = (float) (period - PERIODS[++note]) / 64.f;
 	}
 
 	return period - (delta * finetune);
